refactor(server): extract server
