+++ crates/kickbase/Cargo.toml
@@ -5,16 +5,6 @@ edition = "2021"
 license = "MIT"
 
 [dependencies]
-api = { path = "../api" }
-askama = "0.12.1"
-askama_axum = "0.4.0"
-axum.workspace = true
-reqwest.workspace = true
-serde.workspace = true
-serde_json.workspace = true
-thiserror.workspace = true
+server = { path = "../server" }
 tokio.workspace = true
-tower.workspace = true
-tower-http.workspace = true
 tracing.workspace = true
-tracing-subscriber.workspace = true
