+++ crates/kickbase/src/main.rs
@@ -1,9 +1,6 @@
-mod server;
-mod trace;
-
 #[tokio::main]
 async fn main() {
-  trace::setup_tracing();
+  server::trace::setup_tracing();
 
   if let Err(err) = server::start_server().await {
     tracing::error!("Server encountered a critical error: {}", err);
