+++ crates/server/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "server"
+version = "0.1.1"
+edition = "2021"
+license = "MIT"
+
+[dependencies]
+api = { path = "../api" }
+askama = "0.12.1"
+askama_axum = "0.4.0"
+axum.workspace = true
+reqwest.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+thiserror.workspace = true
+tokio.workspace = true
+tower.workspace = true
+tower-http.workspace = true
+tracing.workspace = true
+tracing-subscriber.workspace = true
