+++ crates/server/assets/htmx-json.js
