+++ crates/server/assets/htmx.js
