+++ crates/server/src/address.rs
@@ -1,4 +1,4 @@
-use crate::server::port::PORT;
+use crate::port::PORT;
 use std::net::SocketAddr;
 
 pub fn get_address() -> SocketAddr {
