+++ crates/server/src/lib.rs
@@ -1,7 +1,8 @@
-mod address;
-mod html;
-mod port;
-mod router;
+pub mod address;
+pub mod html;
+pub mod port;
+pub mod router;
+pub mod trace;
 
 use address::get_address;
 use tokio::net::TcpListener;
