+++ crates/server/src/router/routes/assets/mod.rs
