+++ crates/server/src/router/routes/home/get.rs
@@ -1,11 +1,10 @@
+use crate::html::HtmlTemplate;
 use askama::Template;
 use askama_axum::IntoResponse;
 
-use crate::server::html::HtmlTemplate;
-
 #[derive(Template)]
-#[template(path = "pages/login/get.html")]
-pub struct Html;
+#[template(path = "pages/home.html")]
+pub struct Html {}
 
 pub async fn route() -> impl IntoResponse {
   let template = Html {};
