+++ crates/server/src/router/routes/login/get.rs
@@ -1,10 +1,11 @@
-use crate::server::html::HtmlTemplate;
 use askama::Template;
 use askama_axum::IntoResponse;
 
+use crate::html::HtmlTemplate;
+
 #[derive(Template)]
-#[template(path = "pages/home.html")]
-pub struct Html {}
+#[template(path = "pages/login/get.html")]
+pub struct Html;
 
 pub async fn route() -> impl IntoResponse {
   let template = Html {};
